+++ src/main/resources/templates/lotto.html
@@ -35,8 +35,8 @@
 <form method="POST" action="/winLotto" id="postLotto">
     <div hidden>
         <input type="hidden" value="{{userLotto}}" name="userLotto"/>
-        <input type="hidden" value="{{round}}" name="round"/>
-        <input type="hidden" value="{{lottoRound}}" name="lottoRound"/>
+        <input type="hidden" value="{{roundDto.round}}" name="round"/>
+        <input type="hidden" value="{{roundDto.lottoRound}}" name="lottoRound"/>
     </div>
     <div id="bottom-panel" style="float: right; margin-right: 500px">
         <button type="submit" class="btn btn-secondary">당첨 번호 입력</button>
